+++ TPINT_GRUPO_6_LAB4/src/main/java/frgp/utn/edu/ar/dtos/StockRequest.java
@@ -15,96 +15,48 @@ public class StockRequest {
 	private Articulo articulo;
 	private Date fechaIngreso;
 	private int cantidad;
-	private float precioCompra;
+	private double precioCompra;
 	
 	
 	public int getId() {
 		return id;
 	}
 
-
-
-
-
-
 	public void setId(int id) {
 		this.id = id;
 	}
-
-
-
-
-
 
 	public Articulo getArticulo() {
 		return articulo;
 	}
 
-
-
-
-
-
 	public void setArticulo(Articulo articulo) {
 		this.articulo = articulo;
 	}
-
-
-
-
-
 
 	public Date getFechaIngreso() {
 		return fechaIngreso;
 	}
 
-
-
-
-
-
 	public void setFechaIngreso(Date fechaIngreso) {
 		this.fechaIngreso = fechaIngreso;
 	}
-
-
-
-
-
 
 	public int getCantidad() {
 		return cantidad;
 	}
 
-
-
-
-
-
 	public void setCantidad(int cantidad) {
 		this.cantidad = cantidad;
 	}
 
-
-
-
-
-
-	public float getPrecioCompra() {
+	public double getPrecioCompra() {
 		return precioCompra;
 	}
 
-
-
-
-
-
-	public void setPrecioCompra(float precioCompra) {
+	public void setPrecioCompra(double precioCompra) {
 		this.precioCompra = precioCompra;
 	}
-
-
-
 
 	public Stock construirStock(){
         Stock s = new Stock(articulo, fechaIngreso, cantidad, precioCompra);
